calm before the storm
