+++ src/app/bridge/util/driver.tsx
@@ -28,359 +28,6 @@ export function sbToString(sb: any): any {
 
 }
 
-export function getCATPuzzle(
-  TAILProgramHash: string,
-  innerPuzzle: SExp
-): SExp {
-  return GreenWeb.util.sexp.curry(
-    GreenWeb.util.sexp.fromHex(CAT_MOD),
-    [
-      GreenWeb.util.sexp.bytesToAtom(CAT_MOD_HASH),
-      GreenWeb.util.sexp.bytesToAtom(TAILProgramHash),
-      innerPuzzle
-    ]
-  )
-}
-
-function getCATSolution(
-  innerPuzzleSolution: SExp,
-  lineageProof: SExp | null,
-  prevCoinId: string,
-  thisCoinInfo: any,
-  nextCoinProof: any,
-  prevSubtotal: BigNumberish,
-  extraDelta: BigNumberish
-): SExp {
-  return SExp.to([
-    innerPuzzleSolution,
-    lineageProof ?? SExp.FALSE,
-    GreenWeb.util.sexp.bytesToAtom(prevCoinId),
-    GreenWeb.util.coin.toProgram(thisCoinInfo),
-    GreenWeb.util.coin.toProgram(nextCoinProof),
-    GreenWeb.util.sexp.bytesToAtom(
-      GreenWeb.util.coin.amountToBytes(prevSubtotal)
-    ),
-    GreenWeb.util.sexp.bytesToAtom(
-      GreenWeb.util.coin.amountToBytes(extraDelta)
-    ),
-  ])
-}
-
-export function getWrappedERC20AssetID(sourceChain: Network, erc20ContractAddress: string) {
-  erc20ContractAddress = GreenWeb.util.unhexlify(erc20ContractAddress)!;
-  erc20ContractAddress = "0".repeat(64 - erc20ContractAddress.length) + erc20ContractAddress;
-
-  return GreenWeb.util.sexp.sha256tree(
-    getWrappedTAIL(
-      CHIA_NETWORK.portalLauncherId!,
-      stringToHex(sourceChain.id),
-      GreenWeb.util.unhexlify(sourceChain.erc20BridgeAddress!)!,
-      erc20ContractAddress
-    )
-  );
-}
-
-export function mintCATs(
-  message: any,
-  portalCoinRecord: any,
-  portalParentSpend: any,
-  nonces: any,
-  chains_and_nonces_used_last_spend: [string, string][],
-  offer: string,
-  sig_strings: string[],
-  sig_switches: boolean[],
-  source_chain: string,
-  source_contract: string,
-  portal_receiver_launcher_id: string,
-) {
-  const {
-    nonce,
-    destination,
-    contents
-  } = message;
-  const [ethAssetContract, xchReceiverPh, tokenAmount] = contents;
-  const tokenAmountInt: number = parseInt(tokenAmount, 16);
-  const offer_sb = offerToSpendBundle(offer);
-
-  /* find source coin = coin with OFFER_MOD that will create CAT minter */
-  var source_coin = new GreenWeb.Coin();
-
-  for(var i = 0; i < offer_sb.coinSpends.length; ++i) {
-    const coinSpend = offer_sb.coinSpends[i];
-
-    var [_, conditionsDict, __] = GreenWeb.util.sexp.conditionsDictForSolution(
-      coinSpend.puzzleReveal,
-      coinSpend.solution,
-      GreenWeb.util.sexp.MAX_BLOCK_COST_CLVM
-    );
-    var createCoinConds = conditionsDict!.get("33" as ConditionOpcode) ?? [];
-
-    for(var j = 0; j < createCoinConds.length; ++j) {
-      const cond = createCoinConds[j];
-
-      if(cond.vars[0] === OFFER_MOD_HASH) {
-        source_coin.parentCoinInfo = GreenWeb.util.coin.getName(coinSpend.coin);
-        source_coin.puzzleHash = OFFER_MOD_HASH;
-        source_coin.amount = tokenAmountInt;
-        break;
-      }
-    }
-  }
-
-  /* beign building spend bundle */
-  var coin_spends = offer_sb.coinSpends;
-  const sigs: string[] = [];
-  sigs.push(offer_sb.aggregatedSignature);
-  
-  /* spend portal to create message */
-  const portalCoin = portalCoinRecord.coin;
-
-  const updatePuzzle = getMOfNDelegateDirectPuzzle(
-    CHIA_NETWORK.multisigThreshold!,
-    CHIA_NETWORK.multisigInfos!,
-  );
-  const portalInnerPuzzle = getPortalReceiverInnerPuzzle(
-    portal_receiver_launcher_id,
-    CHIA_NETWORK.signatureThreshold!,
-    CHIA_NETWORK.validatorInfos!,
-    GreenWeb.util.sexp.sha256tree(updatePuzzle),
-    chains_and_nonces_used_last_spend
-  );
-  const portalPuzzle = GreenWeb.util.sexp.singletonPuzzle(portal_receiver_launcher_id, portalInnerPuzzle);
-
-  var portalParentInnerPuzHash = null;
-  const parentPuzzle = GreenWeb.util.sexp.fromHex(
-    GreenWeb.util.unhexlify(portalParentSpend.puzzle_reveal)!
-  );
-  if(GreenWeb.util.sexp.sha256tree(parentPuzzle ) !== SINGLETON_LAUNCHER_HASH) {
-    const [_, args] = GreenWeb.util.sexp.uncurry(parentPuzzle)!;
-    const innerPuzzle = args[1];
-    portalParentInnerPuzHash = GreenWeb.util.sexp.sha256tree(innerPuzzle);
-  }
-  const portalLineageProof = portalParentInnerPuzHash !== null ? SExp.to([
-    Bytes.from(portalParentSpend.coin.parent_coin_info, "hex"),
-    Bytes.from(portalParentInnerPuzHash, "hex"),
-    Bytes.from("01", "hex"),
-  ]) : SExp.to([
-    Bytes.from(portalParentSpend.coin.parent_coin_info, "hex"),
-    Bytes.from("01", "hex"),
-  ]);
-
-  const portalInnerSolution = getPortalReceiverInnerSolution(
-    sig_switches,
-    nonce,
-    source_chain,
-    GreenWeb.util.unhexlify(source_contract)!,
-    destination,
-    contents
-  );
-  const portalSolution = GreenWeb.util.sexp.singletonSolution(
-    portalLineageProof,
-    1,
-    portalInnerSolution
-  );
-
-  const portalCoinSpend = new GreenWeb.util.serializer.types.CoinSpend();
-  portalCoinSpend.coin = GreenWeb.util.goby.parseGobyCoin({
-    amount: 1,
-    parent_coin_info: GreenWeb.util.unhexlify(portalCoin.parent_coin_info),
-    puzzle_hash: GreenWeb.util.unhexlify(portalCoin.puzzle_hash)
-  })!;
-  portalCoinSpend.puzzleReveal = portalPuzzle;
-  portalCoinSpend.solution = portalSolution;
-  coin_spends.push(portalCoinSpend);
-
-  /* spend source coin to create minter */
-  const minterPuzzle = getCATMinterPuzzle(
-    portal_receiver_launcher_id,
-    source_chain,
-    GreenWeb.util.unhexlify(source_contract)!
-  );
-  const minterPuzzleHash = GreenWeb.util.sexp.sha256tree(minterPuzzle);
-  const sourceCoinSolution = SExp.to([
-    [
-      GreenWeb.util.sexp.bytesToAtom(
-        GreenWeb.util.unhexlify(nonce)!
-      ),
-      [
-        GreenWeb.util.sexp.bytesToAtom(minterPuzzleHash),
-        tokenAmountInt
-      ]
-    ],
-  ]);
-
-  const sourceCoinSpend = new GreenWeb.util.serializer.types.CoinSpend();
-  sourceCoinSpend.coin = source_coin;
-  sourceCoinSpend.puzzleReveal = GreenWeb.util.sexp.fromHex(OFFER_MOD);
-  sourceCoinSpend.solution = sourceCoinSolution;
-  coin_spends.push(sourceCoinSpend);
-
-  /* spend minter coin */
-  const minterCoin = new GreenWeb.Coin();
-  minterCoin.parentCoinInfo = GreenWeb.util.coin.getName(source_coin);
-  minterCoin.puzzleHash = minterPuzzleHash;
-  minterCoin.amount = tokenAmountInt;
-
-  const minterSolution = getCATMinterPuzzleSolution(
-    GreenWeb.util.unhexlify(nonce)!,
-    contents,
-    minterCoin.puzzleHash,
-    GreenWeb.util.coin.getName(minterCoin),
-    GreenWeb.util.coin.getName(portalCoinSpend.coin)
-  );
-
-  const minterCoinSpend = new GreenWeb.util.serializer.types.CoinSpend();
-  minterCoinSpend.coin = minterCoin;
-  minterCoinSpend.puzzleReveal = minterPuzzle;
-  minterCoinSpend.solution = minterSolution;
-  coin_spends.push(minterCoinSpend);
-  
-  /* spend message coin */
-  const messageCoinPuzzle = getMessageCoinPuzzle(
-    portal_receiver_launcher_id,
-    source_chain,
-    GreenWeb.util.unhexlify(source_contract)!,
-    GreenWeb.util.unhexlify(nonce)!,
-    destination,
-    GreenWeb.util.sexp.sha256tree(
-      getMessageAsSExp(contents)
-    )
-  );
-  console.log({ messageSExp: GreenWeb.util.sexp.toHex(getMessageAsSExp(contents)) })
-
-  const messageCoin = new GreenWeb.Coin();
-  messageCoin.parentCoinInfo = GreenWeb.util.coin.getName(portalCoinSpend.coin);
-  messageCoin.puzzleHash = GreenWeb.util.sexp.sha256tree(messageCoinPuzzle);
-  messageCoin.amount = 0;
-
-  const messageCoinSolution = getMessageCoinSolution(
-    minterCoin,
-    portalCoin.parent_coin_info,
-    GreenWeb.util.sexp.sha256tree(portalInnerPuzzle),
-    GreenWeb.util.coin.getName(messageCoin)
-  );
-
-  const messageCoinSpend = new GreenWeb.util.serializer.types.CoinSpend();
-  messageCoinSpend.coin = messageCoin;
-  messageCoinSpend.puzzleReveal = messageCoinPuzzle;
-  messageCoinSpend.solution = messageCoinSolution;
-  coin_spends.push(messageCoinSpend);
-
-  /* spend eve CAT coin */
-  const wrappedAssetTAIL = getWrappedTAIL(
-    portal_receiver_launcher_id,
-    source_chain,
-    GreenWeb.util.unhexlify(source_contract)!,
-    GreenWeb.util.unhexlify(ethAssetContract)!
-  );
-  const wrappedAssetTAILHash = GreenWeb.util.sexp.sha256tree(wrappedAssetTAIL);
-
-  const mintAndPayoutInnerPuzzle = getCATMintAndPayoutInnerPuzzle(
-    xchReceiverPh
-  );
-
-  const eveCATPuzzle = getCATPuzzle(
-    wrappedAssetTAILHash,
-    mintAndPayoutInnerPuzzle
-  )
-  const eveCATPuzzleHash = GreenWeb.util.sexp.sha256tree(eveCATPuzzle);
-
-  const eveCAT = new GreenWeb.Coin();
-  eveCAT.parentCoinInfo = GreenWeb.util.coin.getName(minterCoin);
-  console.log({ minterCoin })
-  eveCAT.puzzleHash = eveCATPuzzleHash;
-  eveCAT.amount = tokenAmountInt;
-
-  const eveCATInnerSolution = getCATMintAndPayoutInnerPuzzleSolution(
-    wrappedAssetTAIL,
-    tokenAmountInt,
-    minterCoin.parentCoinInfo
-  );
-  const eveCATProof = {
-    parentCoinInfo: eveCAT.parentCoinInfo,
-    puzzleHash: GreenWeb.util.sexp.sha256tree(mintAndPayoutInnerPuzzle), // inner puzzle hash since this is a CAT proof
-    amount: eveCAT.amount
-  }
-  const eveCATSolution = getCATSolution(
-    eveCATInnerSolution,
-    null,
-    GreenWeb.util.coin.getName(eveCAT),
-    eveCAT,
-    eveCATProof,
-    0,
-    0
-  );
-
-  const eveCATSpend = new GreenWeb.util.serializer.types.CoinSpend();
-  eveCATSpend.coin = eveCAT;
-  eveCATSpend.puzzleReveal = eveCATPuzzle;
-  eveCATSpend.solution = eveCATSolution;
-  coin_spends.push(eveCATSpend);
-
-  /* lastly, aggregate sigs  and build spend bundle */
-
-  sig_strings.map((sig_string) => {
-    var [
-      origin_chain,
-      destination_chain,
-      nonce,
-      coin_id,
-      sig
-    ] = decodeSignature(sig_string);
-    console.log({origin_chain,
-      nonce,
-      coin_id,
-      sig
-    });
-    sigs.push(sig);
-  });
-
-  const { AugSchemeMPL, G2Element } = getBLSModule();
-
-  console.log({ sigs })
-
-  const sb = new GreenWeb.util.serializer.types.SpendBundle();
-  sb.coinSpends = coin_spends;
-  sb.aggregatedSignature = Buffer.from(
-    AugSchemeMPL.aggregate(
-      sigs.map((sig) => G2Element.from_bytes(Buffer.from(sig, "hex")))
-    ).serialize()
-  ).toString("hex");
-  // console.log( sbToString(sb) );
-
-  return {
-    sb,
-    txId: GreenWeb.util.coin.getName(messageCoin)
-  };
-}
-
-export function getBurnSendAddress(
-  destination_chain: string,
-  destination: string,
-  source_chain_token_contract_address: string,
-  target_receiver: string,
-  prefix: string = "xch",
-  bridging_toll_mojos: number
-) {
-  source_chain_token_contract_address = GreenWeb.util.unhexlify(source_chain_token_contract_address)!;
-  source_chain_token_contract_address = "0".repeat(64 - source_chain_token_contract_address.length) + source_chain_token_contract_address;
-
-  const burnInnerPuzzle = getCATBurnInnerPuzzle(
-    destination_chain,
-    destination,
-    source_chain_token_contract_address,
-    target_receiver,
-    bridging_toll_mojos
-  );
-
-  const burnInnerPuzzleHash = GreenWeb.util.sexp.sha256tree(burnInnerPuzzle);
-
-  return GreenWeb.util.address.puzzleHashToAddress(
-    burnInnerPuzzleHash,
-    prefix
-  );
-}
-
 export function getBurnSendFullPuzzleHash(
   destination_chain: string,
   destination: string,
