+++ src/app/bridge/util/sig.tsx
@@ -41,20 +41,6 @@ export function decodeSignature(sig: string): [
   return [originChain, destinationChain, nonce, coinId, sigData];
 }
 
-export function stringToHex(str: string): string {
-    return str.split('').map(c => c.charCodeAt(0).toString(16).padStart(2, '0')).join('');
-}
-
-
-export function hexToString(hex: string): string {
-    let str = '';
-    for (let i = 0; i < hex.length; i += 2) {
-        str += String.fromCharCode(parseInt(hex.substring(i, i + 2), 16));
-    }
-    return str;
-}
-
-
 export async function getSigsAndSelectors(
   sourceChainHex: string,
   destinationChainHex: string,
